bridegif_ns: fix idempotency (#80) (#81)
bridgeif_ns was working on the erroneous assumption that
creating/deleting one end of the link from the main name space
into the project name space would take care of the other end
as well. That is not the cause and caused numerous problems
with the build and read verbs. This commit addresses the issue
by checking for and handling dangling interfaces with one side
missing.