hyperv: fix braces in quiesce() (#95)
We had overzealous escaping of curly braces in a string that wasn't an
f-string. That lead to the shutdown_vm payload incorrectly having double
curly braces around VM state check loop's body. With these double curly
braces it never terminates if the payload is run with an invalid VM
name.